Refactor MiniDNS
