+++ minidns-core/src/main/java/de/measite/minidns/dnsserverlookup/DNSServerLookupMechanism.java
@@ -1,0 +1,11 @@
+package de.measite.minidns.dnsserverlookup;
+
+public interface DNSServerLookupMechanism extends Comparable<DNSServerLookupMechanism> {
+
+    public String getName();
+
+    public int getPriority();
+
+    public String[] getDnsServerAddresses();
+
+}
